scripts/requirements.txt
# AI Image Generation Dependencies
google-genai>=0.2.0
pillow>=10.0.0
rembg>=2.0.50
